refactoring: create middleware concept, make session handling a middleware, add backends to handle starting the PS process
The new backends make it possible to not just run a PS process
on the local machine, but also use things like SSH connections
to start remote processes. Even though "powershell over ssh"
is probably a somewhat exotic topic.

The session stuff was basically always just a wrapper around
the underlying shell, and when the need for encoding handling
arose, it became obvious that the UTF8 decorator is basically
the same as the session decorator, so now we have a middleware
namespace where these things live.

Another change in this commit is that closing a session also
closes the underlying shell process. This is due to the new
middleware concept, where a user of a shell does not want to
know how many underlying middlewares there are, so it makes
sense to close the underlying middleware as well. It's
technically still possible to use the session along "regular"
commands, developers need justt ake care of properly handling
the .Exit() call.